Fixed errors Middleware Typing
The `errors` function is *not* a middleware, but rather *returns* a middleware.  As a result, when adding `errors` as a middleware, you're unable to call it and forced to just provide the function itself, which in turn creates an infinite loop as the middleware factory never calls `next`.

This change fixes that issue.